refactor: Companion Object Factory Functions 으로 변경
기존 LottoVendingMachine에서 Lottos의 Companion Object로 이동
