+++ src/main/kotlin/lotto/application/LottoApplication.kt
@@ -2,8 +2,8 @@ package lotto.application
 
 import lotto.domain.Lotto
 import lotto.domain.LottoResult
-import lotto.LottoVendingMachine
 import lotto.domain.LottoNumber
+import lotto.domain.Lottos
 import lotto.view.InputView
 import lotto.view.OutputView
 
@@ -14,14 +14,14 @@ class LottoApplication(
 
     fun run() {
         val money = inputView.inputMoney()
-        val lotto = LottoVendingMachine.buyLotto(money)
+        val lotto = Lottos.buyLotto(money)
         outputView.showLottoCount(lotto)
         outputView.showLotto(lotto)
 
         val winningLotto: Lotto = inputView.inputWinningNumbers()
         val bonusNumber: LottoNumber = inputView.inputBonusNumber()
 
-        val lottoResult= LottoResult.makeLottoResult(
+        val lottoResult = LottoResult.makeLottoResult(
             winningLotto = winningLotto,
             bonusLottoNumber = bonusNumber,
             lottos = lotto,
